+++ cocos/base/CCString.cpp
@@ -221,7 +221,7 @@ String* String::create(const std::string& str)
     return ret;
 }
 
-String* String::createWithData(const unsigned char* data, unsigned long nLen)
+String* String::createWithData(const unsigned char* data, int nLen)
 {
     String* ret = NULL;
     if (data != NULL)
@@ -255,11 +255,11 @@ String* String::createWithFormat(const char* format, ...)
 
 String* String::createWithContentsOfFile(const char* filename)
 {
-    long size = 0;
+    ssize_t size = 0;
     unsigned char* data = 0;
     String* ret = NULL;
     data = FileUtils::getInstance()->getFileData(filename, "rb", &size);
-    ret = String::createWithData(data, size);
+    ret = String::createWithData(data, static_cast<int>(size));
     free(data);
     return ret;
 }
